Enhance chains configuration loading and introduce week start parsing
This update refactors the `loadChainsConfig` function to streamline error handling and improve clarity by throwing errors for failed fetch requests. Additionally, a new `parseWeekStart` function has been added to the `utils.js` file, allowing for better handling of week start values across the application. The `maintenance.js` and `chains.js` files have been updated to utilize this new function, ensuring consistent week start parsing and improving overall code maintainability.
